Refactor chatHandler.js god-module into focused utils/chat/ modules
chatHandler.js (2101 lines) split into:
- utils/chat/chatDb.js: user/conversation rows, message tracking, system-log
  and health diagnostics
- utils/chat/chatContext.js: context-window assembly + rolling summaries
- utils/chat/searchFlow.js: legacy search detection/approval flow and
  /search + /generate response directives
- utils/chat/reactions.js: reaction controls (regenerate, pin, branch,
  deep dive, summarize)
- utils/chat/responder.js: chunked reply delivery
- utils/chat/threadManager.js: race-free thread creation + naming
- utils/chat/prompts.js: built-in default system prompt

chatHandler.js keeps the orchestration (handleChatInteraction,
processMessage) and re-exports the same public API, so all existing
require() sites are unchanged. Function bodies moved verbatim except:
dead variables removed (unused thread-preference lookup, unused constants),
and the unused pendingRequests import cleanup from the previous commit.
