+++ utils/chatHandler.js
@@ -1,907 +1,49 @@
+/**
+ * Chat pipeline orchestrator. This module owns the main interaction flow
+ * (handleChatInteraction / processMessage); the supporting pieces live in
+ * focused modules under utils/chat/:
+ *   - chatDb.js        user/conversation rows, message tracking, diagnostics
+ *   - chatContext.js   context window assembly + rolling summaries
+ *   - searchFlow.js    legacy search detection/approval + response directives
+ *   - reactions.js     reaction controls on bot replies
+ *   - responder.js     chunked reply delivery
+ *   - threadManager.js race-free thread creation + thread naming
+ *   - prompts.js       built-in default system prompt
+ *
+ * The public API re-exported here is unchanged so existing require() sites
+ * keep working.
+ */
 const db = require('../db');
-const { ThreadAutoArchiveDuration } = require('discord.js');
-const AISearchHandler = require('./aiSearchHandler');
 const { chunkMessage } = require('./index');
-const { getPrompt, getPromptWithGuildPersonality } = require('./memeMode');
-const { getThreadPreference, THREAD_PREFERENCE, getPersonalityDirective, getGuildAI } = require('./guildSettings');
+const { getPersonalityDirective, getGuildAI } = require('./guildSettings');
 const aiService = require('../services/aiService');
 const imageDetectionHandler = require('./imageDetectionHandler');
 const path = require('path');
-const { setInterval } = require('timers');
 const { getGuildContext, getPreferredUserName, getBotPreferredName } = require('./guildContext');
 const toolsRegistry = require('./toolsRegistry');
 const memoryService = require('../services/memoryService');
 const factsService = require('../services/factsService');
 
-// Add a Map to track pending image generations by channel
-const pendingImageGenerations = new Map();
-
-// Add at the top with other constants
-const threadLocks = new Map();
-
-// Add near the top with other constants
-const MAX_WORD_LENGTH = 500;
-
-// Add near the top with other constants
-const TRANSACTION_TIMEOUT = 30000; // 30 seconds
-
-// Add a global flag to track database connectivity state
-let dbConnectivityOK = true;
-let lastDbHealthCheck = Date.now();
-const DB_HEALTH_CHECK_INTERVAL = 10 * 60 * 1000; // 10 minutes
-
-/**
- * Logs a system event to the database
- * @param {string} level - The log level (ERROR, WARN, INFO, DEBUG)
- * @param {string} message - The log message
- * @param {Object} metadata - Additional metadata to store
- * @returns {Promise<void>}
- */
-async function logSystemEvent(level, message, metadata = {}) {
-    try {
-        db.run(
-            `INSERT INTO system_logs (log_level, message, metadata, source, error_code, error_state)
-             VALUES (@level, @message, @metadata, @source, @errorCode, @errorState)`,
-            {
-                level,
-                message,
-                metadata: JSON.stringify(metadata),
-                source: metadata.source || null,
-                errorCode: metadata.error_code || null,
-                errorState: metadata.error_state || null
-            }
-        );
-    } catch (error) {
-        console.error('Error logging system event:', error);
-    }
-}
-
-// Schedule periodic database health checks
-setInterval(checkDatabaseHealth, DB_HEALTH_CHECK_INTERVAL).unref?.();
-
-/**
- * Looks up a user by Discord ID, creating the record if needed.
- * @param {string} discordId - The Discord user ID (snowflake)
- * @param {string} username - Username to store when creating the record
- * @returns {number} Internal user id
- */
-function getOrCreateUser(discordId, username) {
-    const existing = db.get('SELECT id FROM users WHERE discordId = @discordId', { discordId });
-    if (existing) return existing.id;
-
-    const result = db.run(
-        'INSERT INTO users (discordUsername, discordId, username) VALUES (@username, @discordId, @username)',
-        { discordId, username }
-    );
-    return Number(result.lastInsertRowid);
-}
-
-/**
- * Looks up a conversation for a user within a guild conversation, creating it if needed.
- * @param {number} userId - Internal user id
- * @param {number} guildConvId - guild_conversations id
- * @returns {number} Conversation id
- */
-function getOrCreateConversation(userId, guildConvId) {
-    const existing = db.get(
-        'SELECT id FROM conversations WHERE userId = @userId AND guildConversationId = @guildConvId',
-        { userId, guildConvId }
-    );
-    if (existing) return existing.id;
-
-    const result = db.run(
-        'INSERT INTO conversations (userId, guildConversationId) VALUES (@userId, @guildConvId)',
-        { userId, guildConvId }
-    );
-    return Number(result.lastInsertRowid);
-}
-
-/**
- * Checks database health by performing basic query operations
- * @returns {Promise<boolean>} True if database is healthy
- */
-async function checkDatabaseHealth() {
-    console.log('Performing database health check...');
-    lastDbHealthCheck = Date.now();
-
-    try {
-        const userCount = db.get('SELECT COUNT(*) as count FROM users').count;
-        const messageCount = db.get('SELECT COUNT(*) as count FROM messages').count;
-
-        // Verify write access with a test insert that is rolled back with the transaction helper.
-        db.transaction(() => {
-            const result = db.run(
-                `INSERT INTO system_logs (log_level, message, source)
-                 VALUES ('DEBUG', 'DB health check - write test', 'checkDatabaseHealth')`
-            );
-            db.run('DELETE FROM system_logs WHERE id = @id', { id: Number(result.lastInsertRowid) });
-        });
-
-        console.log('Database health check successful', {
-            userCount,
-            messageCount,
-            time: new Date().toISOString()
-        });
-
-        dbConnectivityOK = true;
-        return true;
-    } catch (error) {
-        dbConnectivityOK = false;
-        console.error('Database health check failed:', {
-            error: error.message,
-            code: error.code,
-            timestamp: new Date().toISOString()
-        });
-        return false;
-    }
-}
-
-/**
- * Diagnoses database connection issues with detailed reporting
- * @param {Object} interaction - The Discord interaction object
- * @returns {Promise<string>} Diagnostic message
- */
-async function diagnoseDatabaseIssues(interaction) {
-    try {
-        console.log('Running database diagnostics...');
-
-        let hasReadPermission = true;
-        let hasWritePermission = true;
-        const detailedErrors = [];
-
-        try {
-            db.get('SELECT * FROM users LIMIT 1');
-        } catch (error) {
-            hasReadPermission = false;
-            detailedErrors.push(`Read Error: ${error.message}`);
-        }
-
-        try {
-            db.transaction(() => {
-                const result = db.run(
-                    `INSERT INTO system_logs (log_level, message, source)
-                     VALUES ('DEBUG', 'DB diagnostics - write test', 'diagnoseDatabaseIssues')`
-                );
-                db.run('DELETE FROM system_logs WHERE id = @id', { id: Number(result.lastInsertRowid) });
-            });
-        } catch (error) {
-            hasWritePermission = false;
-            detailedErrors.push(`Write Error: ${error.message}`);
-        }
-
-        let diagnosticMessage = "**Database Diagnostic Results**\n";
-
-        if (hasReadPermission && hasWritePermission) {
-            diagnosticMessage += "✅ Database connection and permissions appear to be working correctly.\n";
-
-            const recentMessageCount = db.get(
-                "SELECT COUNT(*) as count FROM messages WHERE createdAt > datetime('now', '-1 day')"
-            );
-            const totalMessageCount = db.get('SELECT COUNT(*) as count FROM messages');
-            const mostRecentMessage = db.get(
-                'SELECT id, createdAt as timestamp, isBot FROM messages ORDER BY createdAt DESC LIMIT 1'
-            );
-
-            diagnosticMessage += `✅ Found ${recentMessageCount.count} messages stored in the last 24 hours.\n`;
-            diagnosticMessage += `✅ Total message count in database: ${totalMessageCount.count}\n`;
-
-            if (mostRecentMessage) {
-                diagnosticMessage += `✅ Most recent message (ID: ${mostRecentMessage.id}) was stored at ${mostRecentMessage.timestamp} (${mostRecentMessage.isBot ? 'bot' : 'user'} message)\n`;
-            } else {
-                diagnosticMessage += `⚠️ No messages found in the database.\n`;
-            }
-
-            // Check for recent errors in system logs
-            const recentErrors = db.all(
-                `SELECT id, createdAt as timestamp, message
-                 FROM system_logs
-                 WHERE log_level = 'ERROR' AND createdAt > datetime('now', '-1 day')
-                 ORDER BY createdAt DESC LIMIT 10`
-            );
-
-            if (recentErrors.length > 0) {
-                diagnosticMessage += `\n**Recent Errors (Last 24h):**\n`;
-                recentErrors.forEach(error => {
-                    diagnosticMessage += `- ${error.timestamp}: ${error.message.substring(0, 100)}...\n`;
-                });
-            } else {
-                diagnosticMessage += "\n**Recent Errors:** No errors logged in the last 24 hours.\n";
-            }
-
-            // Database file statistics
-            const pageCount = db.getDb().pragma('page_count', { simple: true });
-            const pageSize = db.getDb().pragma('page_size', { simple: true });
-            const dbSizeMb = ((pageCount * pageSize) / (1024 * 1024)).toFixed(2);
-            diagnosticMessage += `\n**Database Info:**\n- Engine: SQLite (better-sqlite3)\n- Size: ${dbSizeMb} MB\n- Journal mode: ${db.getDb().pragma('journal_mode', { simple: true })}\n`;
-        } else {
-            if (!hasReadPermission) {
-                diagnosticMessage += "❌ Cannot read from database tables. This may be a permissions issue.\n";
-            }
-
-            if (!hasWritePermission) {
-                diagnosticMessage += "❌ Cannot write to database tables. This may be a permissions issue.\n";
-            }
-
-            diagnosticMessage += "\nDetailed Errors:\n";
-            detailedErrors.forEach(error => {
-                diagnosticMessage += `- ${error}\n`;
-            });
-        }
-
-        return diagnosticMessage;
-
-    } catch (error) {
-        console.error('Error in database diagnosis:', error);
-        return `Failed to complete database diagnosis: ${error.message}`;
-    }
-}
-
-/**
- * Creates a placeholder thread ID for channel-only conversations
- * @param {string} channelId - The Discord channel ID
- * @returns {string} - A placeholder thread ID
- */
-function createPlaceholderThreadId(channelId) {
-    return `channel-${channelId}`;
-}
-
-// In-flight search deduplication, persisted in SQLite (pending_searches) so
-// a restart between detection and completion cannot double-fire a search.
-function isSearchPending(channelId, query) {
-    try {
-        // Clean up old searches (older than 5 minutes)
-        db.run(`DELETE FROM pending_searches WHERE createdAt < datetime('now', '-5 minutes')`);
-        return Boolean(db.get(
-            'SELECT 1 FROM pending_searches WHERE channelId = @channelId AND query = @query',
-            { channelId, query }
-        ));
-    } catch (error) {
-        console.error('Error checking pending search:', error.message);
-        return false;
-    }
-}
-
-// Add function to track a new search
-function trackSearch(channelId, query) {
-    try {
-        db.run(
-            `INSERT INTO pending_searches (channelId, query) VALUES (@channelId, @query)
-             ON CONFLICT(channelId, query) DO UPDATE SET createdAt = CURRENT_TIMESTAMP`,
-            { channelId, query }
-        );
-    } catch (error) {
-        console.error('Error tracking pending search:', error.message);
-    }
-}
-
-// Add function to remove a completed search
-function completeSearch(channelId, query) {
-    try {
-        db.run(
-            'DELETE FROM pending_searches WHERE channelId = @channelId AND query = @query',
-            { channelId, query }
-        );
-    } catch (error) {
-        console.error('Error completing pending search:', error.message);
-    }
-}
-
-const DEFAULT_PROMPT = `You are Goobster, a quirky and clever Discord bot with a passion for helping users and a dash of playful sass. You love making witty observations and dropping the occasional pun, but you always stay focused on being genuinely helpful.
-
-Key Traits:
-- Friendly and approachable, but not afraid to show personality
-- Loves making clever wordplay and references when appropriate
-- Takes pride in being accurate and helpful
-- Excited about learning new things alongside users
-
-You have access to real-time web search capabilities through the /search command. When users ask for current information or facts you're not certain about, you should:
-
-1. Acknowledge their request 
-2. Use the /search command by replying with a message in this EXACT format (including quotes):
-   "/search query:"your search query here" reason:"why you need this information""
-
-You also have image generation capabilities! When users ask you to create, draw, or generate an image, you can:
-
-1. Acknowledge their request
-2. Use the built-in image generation by replying with a message in this EXACT format (including quotes):
-   "/generate image:"detailed description of what to generate" type:"CHARACTER|SCENE|LOCATION|ITEM" style:"fantasy|realistic|anime|comic|watercolor|oil_painting""
-
-Example image generation responses:
-
-For character portraits:
-"I'd love to visualize that character for you! /generate image:"tall elven warrior with silver hair and emerald eyes, wearing ornate plate armor with flowing blue cape" type:"CHARACTER" style:"fantasy""
-
-For scenes:
-"Let me create that scene! /generate image:"futuristic cyberpunk city street at night with neon signs and flying cars" type:"SCENE" style:"realistic""
-
-For locations:
-"I'll draw that place for you! /generate image:"ancient stone temple ruins in a dense jungle with vines and statues" type:"LOCATION" style:"watercolor""
-
-For items:
-"Let me show you how I imagine that! /generate image:"ornate magical staff with glowing crystal and dragon motifs" type:"ITEM" style:"fantasy""
-
-Example search responses:
-
-When needing current info:
-"Let me check the latest data on that! /search query:"current cryptocurrency market trends March 2024" reason:"User asked about crypto prices, and even a bot as clever as me needs up-to-date numbers to give accurate advice!""
-
-When verifying facts:
-"I want to make sure I give you the most accurate info! /search query:"latest Mars rover discoveries 2024" reason:"Need to verify recent Mars exploration data""
-
-Remember:
-- Be enthusiastic but professional
-- Make search queries and image prompts specific and focused
-- Use appropriate emojis and formatting to make responses engaging
-- Stay helpful and informative while maintaining your quirky personality`;
-
-const CONTEXT_WINDOW_SIZE = 20; // Number of messages to keep in active context
-const SUMMARY_TRIGGER = 30; // Number of messages that triggers a summary
-
-async function summarizeContext(messages, guildConvId) {
-    try {
-        const messageText = messages
-            .map(m => `${m.role}: ${m.content}`)
-            .join('\n');
-
-        const summaryPrompt = `Please provide a brief, bullet-point summary of the key points from this conversation. Focus on the most important information that would be relevant for future context:\n\n${messageText}`;
-
-        const summary = await aiService.chatText([
-            { role: 'user', content: summaryPrompt }
-        ], {
-            temperature: 0.7,
-            max_tokens: 500
-        });
-        
-        // Chunk the summary if needed
-        const chunks = chunkMessage(summary);
-
-        // Store the summary
-        try {
-            db.run(
-                `INSERT INTO conversation_summaries (guildConversationId, summary, messageCount)
-                 VALUES (@guildConvId, @summary, @messageCount)`,
-                { guildConvId, summary: chunks[0], messageCount: messages.length }
-            );
-        } catch (dbError) {
-            console.error('Database Error:', dbError);
-            throw new Error('Failed to store conversation summary in database.', { cause: dbError });
-        }
-
-        return chunks[0]; // Use first chunk as summary
-    } catch (error) {
-        console.error('Error summarizing context:', error);
-        throw error;
-    }
-}
-
-async function getContextWithSummary(thread, guildConvId, userId = null, interaction = null) {
-    // Get recent messages
-    let messages;
-    let botUserId;
-    let guildId = null;
-    
-    if (thread) {
-        // If we have a thread, fetch messages from it
-        messages = await thread.messages.fetch({ limit: CONTEXT_WINDOW_SIZE });
-        botUserId = thread.client.user.id;
-        guildId = thread.guild?.id;
-    } else if (interaction && interaction.channel) {
-        // If we don't have a thread but have a channel, fetch messages from the channel
-        messages = await interaction.channel.messages.fetch({ limit: CONTEXT_WINDOW_SIZE });
-        botUserId = interaction.client.user.id;
-        guildId = interaction.guild?.id;
-    } else {
-        // If we have neither thread nor channel, return an empty array
-        // The system prompt will be added by the calling function
-        return [];
-    }
-    
-    const conversationHistory = messages
-        .reverse()
-        .map(m => {
-            const isBot = m.author.id === botUserId;
-            const speakerName = isBot ? 'Goobster' : (m.member?.displayName || m.author.username || 'Unknown');
-
-            // Pre-pend the speaker name for clarity when not the bot
-            const contentPrefix = isBot ? '' : `${speakerName}: `;
-
-            return {
-                role: isBot ? 'assistant' : 'user',
-                content: `${contentPrefix}${m.content}`.trim(),
-                messageId: m.id,
-                authorId: m.author.id
-            };
-        })
-        .filter(m => m.content && !m.content.startsWith('/'));
-
-    // If user-specific context is requested, prioritize their messages
-    if (userId) {
-        conversationHistory.sort((a, b) => {
-            if (a.authorId === userId && b.authorId !== userId) return -1;
-            if (a.authorId !== userId && b.authorId === userId) return 1;
-            return 0;
-        });
-    }
-
-    // Handle message references and quotes
-    for (let i = 0; i < conversationHistory.length; i++) {
-        const msg = messages.find(m => m.id === conversationHistory[i].messageId);
-        if (msg?.reference?.messageId) {
-            const referencedMsg = messages.find(m => m.id === msg.reference.messageId);
-            if (referencedMsg) {
-                conversationHistory[i].content = `[Replying to: "${referencedMsg.content.substring(0, 50)}${referencedMsg.content.length > 50 ? '...' : ''}"]\n${conversationHistory[i].content}`;
-            }
-        }
-    }
-
-    // Don't add system prompt here - will be handled by the main chat handler
-    // to ensure personality directive is applied correctly
-
-    // Check if we need to generate a summary
-    if (messages.size >= SUMMARY_TRIGGER) {
-        const summaryRow = db.get(
-            `SELECT summary FROM conversation_summaries
-             WHERE guildConversationId = @guildConvId
-             ORDER BY createdAt DESC LIMIT 1`,
-            { guildConvId }
-        );
-
-        if (summaryRow) {
-            // Add summary as a system message at the beginning
-            conversationHistory.unshift({
-                role: 'system',
-                content: `Previous conversation summary:\n${summaryRow.summary}`
-            });
-        } else {
-            const summary = await summarizeContext(conversationHistory, guildConvId);
-            if (summary) {
-                // Add generated summary as a system message
-                conversationHistory.unshift({
-                    role: 'system',
-                    content: `Previous conversation summary:\n${summary}`
-                });
-            }
-        }
-    }
-
-    return conversationHistory;
-}
-
-// Add function to check for existing search requests
-async function checkExistingSearchRequest(channel, query, botId) {
-    try {
-        const recentMessages = await channel.messages.fetch({ limit: 5 });
-        return recentMessages.find(msg => 
-            msg.author.id === botId && 
-            msg.content.includes('Search Request') &&
-            msg.content.includes(query)
-        );
-    } catch (error) {
-        console.error('Error checking for existing search request:', {
-            error: error.message || 'Unknown error',
-            stack: error.stack || 'No stack trace available',
-            channelId: channel.id,
-            query
-        });
-        return null;
-    }
-}
-
-async function handleAIResponse(response, interaction) {
-    // Add null/undefined check at the beginning
-    if (!response || typeof response !== 'string') {
-        console.warn('handleAIResponse received null/undefined response:', response);
-        return response; // Return as-is if not a string
-    }
-    
-    // Check for search request
-    const searchMatch = response.match(/\/search query:"([^"]+)"\s+reason:"([^"]+)"/);
-    if (searchMatch) {
-        const [, query, reason] = searchMatch;
-        try {
-            // Use consolidated check for existing search request
-            const existingRequest = await checkExistingSearchRequest(
-                interaction.channel,
-                query,
-                interaction.client.user.id
-            );
-
-            if (existingRequest) {
-                return `I've already requested a search for "${query}". Please approve or deny that request first.`;
-            }
-
-            // Request search permission
-            const requestData = await AISearchHandler.requestSearch(interaction, query, reason);
-            if (requestData && requestData.requestId === null) {
-                const guildId = interaction.guild?.id;
-                const systemPrompt = await getPromptWithGuildPersonality(interaction.user.id, guildId);
-                const response = await aiService.chatText([
-                        { role: 'system', content: systemPrompt },
-                        { role: 'user', content: `I need to search for "${query}". ${reason ? `Reason: ${reason}` : ''}` },
-                        { role: 'system', content: `Here is relevant information: ${requestData.result}` }
-                    ], {
-                        preset: 'chat',
-                        max_tokens: 150
-                    });
-
-                const chunks = chunkMessage(response);
-                await interaction.editReply({ content: chunks[0], ephemeral: true });
-                for (let i = 1; i < chunks.length; i++) {
-                    await interaction.followUp({ content: chunks[i], ephemeral: true });
-                }
-                return null;
-            }
-            return `🔍 I've requested permission to search for information about "${query}". Please approve or deny the request.`;
-        } catch (error) {
-            console.error('Error requesting search permission:', {
-                error: error.message || 'Unknown error',
-                stack: error.stack || 'No stack trace available',
-                query,
-                reason
-            });
-            return `I apologize, but I encountered an error while trying to search for information. Let me try to help without the search.`;
-        }
-    }
-
-    // Check for image generation intent
-    const imageMatch = response.match(/\/generate image:"([^"]+)"\s+type:"([^"]+)"\s+style:"([^"]*)"/i);
-    if (imageMatch) {
-        const [, prompt, type, style] = imageMatch;
-        try {
-            // Check if we're already generating an image for this channel
-            const channelId = interaction.channelId;
-            if (pendingImageGenerations.has(channelId)) {
-                return `I'm already working on generating an image. Please wait for that to complete first.`;
-            }
-
-            // Mark that we're generating an image
-            pendingImageGenerations.set(channelId, Date.now());
-
-            // Generate the image
-            const imageType = type.toUpperCase();
-            const validTypes = ['CHARACTER', 'SCENE', 'LOCATION', 'ITEM'];
-            
-            const finalType = validTypes.includes(imageType) ? imageType : 'SCENE';
-            const finalStyle = style || 'fantasy';
-
-            // Send a message indicating we're generating an image
-            await interaction.editReply(`🎨 I'm generating an image of: ${prompt}\nThis might take a moment...`);
-
-            // Generate the image
-            const imagePath = await imageDetectionHandler.generateImage(prompt, finalType, finalStyle);
-            
-            // Clear the pending flag
-            pendingImageGenerations.delete(channelId);
-
-            // Return a special marker for the processMessage function to handle
-            return `__IMAGE_GENERATION_RESULT__${imagePath}__END_IMAGE_GENERATION__`;
-        } catch (error) {
-            // Clear the pending flag on error
-            pendingImageGenerations.delete(interaction.channelId);
-            
-            console.error('Error generating image:', error);
-            return `I apologize, but I encountered an error while trying to generate that image: ${error.message}`;
-        }
-    }
-
-    return response;
-}
-
-/**
- * Uses AI to detect if a message requires a search
- * @param {string} message - The user's message
- * @returns {Promise<Object>} - Object with needsSearch flag and suggested query if needed
- */
-async function detectSearchNeed(message) {
-    try {
-        // First, use a simple prompt to determine if the message needs a search
-        const searchDetectionPrompt = `
-You are an AI assistant that determines if a user message requires a web search to provide an accurate response.
-
-User message: "${message}"
-
-Analyze the message and determine if it:
-1. Asks for current events, news, or time-sensitive information
-2. Requests factual information that might change over time
-3. Asks about specific data, statistics, or facts that you might not have in your training data
-4. Explicitly asks to search for something
-
-Respond with ONLY "true" if a web search is needed, or "false" if no search is needed.
-`;
-
-        const needsSearchResponse = await aiService.generateText(searchDetectionPrompt, {
-            temperature: 0.1, // Low temperature for more deterministic response
-            max_tokens: 10,   // Very short response needed
-        });
-
-        const needsSearch = needsSearchResponse.trim().toLowerCase() === 'true';
-
-        if (needsSearch) {
-            // If search is needed, use a second prompt to generate the optimal search query
-            const searchQueryPrompt = `
-You are an AI assistant that creates effective search queries based on user messages.
-
-User message: "${message}"
-
-Create a concise, specific search query that will find the most relevant information to answer this message.
-Focus on the key information need and remove any unnecessary context or pleasantries.
-Make sure to create a query that will find the MOST CURRENT information available.
-DO NOT include specific years in the query unless the user explicitly asked about a specific year.
-Respond with ONLY the search query text, nothing else.
-`;
-
-            const suggestedQuery = await aiService.generateText(searchQueryPrompt, {
-                temperature: 0.3,
-                max_tokens: 50,
-                includeCurrentDate: true // Include current date in the prompt
-            });
-
-            return {
-                needsSearch: true,
-                suggestedQuery: suggestedQuery.trim(),
-                reason: `User asked about information that may require a search: ${message}`
-            };
-        }
-
-        return { needsSearch: false };
-    } catch (error) {
-        console.error('Error in AI-based search detection:', error);
-        
-        // Fall back to the original keyword-based detection if AI detection fails
-        return fallbackDetectSearchNeed(message);
-    }
-}
-
-/**
- * Fallback method using keyword matching to detect search needs
- * @param {string} message - The user's message
- * @returns {Object} - Object with needsSearch flag and suggested query if needed
- */
-function fallbackDetectSearchNeed(message) {
-    const searchIndicators = [
-        /current|latest|recent|news|today|yesterday|this week|this month|this year/i,
-        /look up|search|find|check|get/i,
-        /what is|who is|where is|when is|why is|how is/i,
-        /what are|who are|where are|when are|why are|how are/i,
-    ];
-
-    // Check if the message contains time-sensitive or search-related keywords
-    const needsSearch = searchIndicators.some(pattern => pattern.test(message));
-
-    if (needsSearch) {
-        // Extract the main topic for search by looking for key phrases
-        let searchTopic = '';
-
-        // Common patterns for search requests
-        const searchPatterns = [
-            /(?:can you |could you |please |)(?:search|look up|find)(?: for| about|) (.*?)(?:\.|\?|!|$)/i,  // "search for X"
-            /(?:what|who|where|when|why|how) (?:is|are|was|were|do|does|did) (.*?)(?:\.|\?|!|$)/i,  // "what is X"
-            /(?:tell me about|find information on|get details on) (.*?)(?:\.|\?|!|$)/i,  // "tell me about X"
-            /(?:I want to know|I need to know|I'd like to know) (?:about |more about |)(.*?)(?:\.|\?|!|$)/i,  // "I want to know about X"
-            /(?:what's|who's|where's|when's|why's|how's) (.*?)(?:\.|\?|!|$)/i,  // "what's X"
-        ];
-
-        for (const pattern of searchPatterns) {
-            const match = message.match(pattern);
-            if (match && match[1]) {
-                searchTopic = match[1].trim();
-                break;
-            }
-        }
-
-        // If no pattern matched, try to extract based on trigger words
-        if (!searchTopic) {
-            const triggerWords = ['look up', 'search', 'find', 'check', 'get', 'latest', 'current'];
-            for (const word of triggerWords) {
-                const index = message.toLowerCase().indexOf(word);
-                if (index !== -1) {
-                    searchTopic = message.slice(index + word.length).split(/[,.!?]/)[0].trim();
-                    break;
-                }
-            }
-        }
-
-        // If we found a topic, return the search info
-        if (searchTopic) {
-            return {
-                needsSearch: true,
-                suggestedQuery: searchTopic,
-                reason: `User asked about current information regarding: ${searchTopic}`
-            };
-        }
-    }
-
-    return { needsSearch: false };
-}
-
-// Handle the search flow. The trimmedMessage parameter contains the user's
-// original message and is used when generating the immediate response after an
-// auto-approved search.
-async function handleSearchFlow(searchInfo, interaction, thread, trimmedMessage) {
-    const channelId = thread?.id || interaction.channel.id;
-    
-    // Check if this search is already pending
-    if (isSearchPending(channelId, searchInfo.suggestedQuery)) {
-        return `I'm already processing a search for "${searchInfo.suggestedQuery}". Please wait for that to complete first.`;
-    }
-    
-    // Track this new search
-    trackSearch(channelId, searchInfo.suggestedQuery);
-    
-    try {
-        // If thread is not provided, try to get it from the interaction
-        if (!thread && interaction.channel && !interaction.channel.isThread() && interaction.guildId) {
-            // Get the guild's thread preference
-            const threadPreference = await getThreadPreference(interaction.guildId);
-            
-            // If preference is ALWAYS_THREAD, create/use a thread
-            if (threadPreference === THREAD_PREFERENCE.ALWAYS_THREAD) {
-                const channelName = interaction.channel.name.replace(/[^a-z0-9-_]/gi, '-').toLowerCase();
-                const threadName = `goobster-chat-${channelName}`;
-                
-                try {
-                    thread = await getOrCreateThreadSafely(interaction.channel, threadName);
-                    
-                    // Send welcome message only for newly created threads
-                    if (!thread.messages.cache.size) {
-                        await thread.send(
-                            "👋 Hi! I've created this thread for our conversation. " +
-                            "You can continue chatting with me here by:\n" +
-                            "1. Using `/chat` command\n" +
-                            `2. Mentioning me (@${interaction.client.user.username})\n\n` +
-                            "The thread will keep our conversation organized and maintain context!"
-                        );
-                    }
-                    
-                    // Notify the user that we've created a thread
-                    if (interaction.channel !== thread) {
-                        await interaction.channel.send({
-                            content: `I've continued our conversation in a thread: ${thread}`,
-                            allowedMentions: { users: [], roles: [] }
-                        });
-                    }
-                } catch (error) {
-                    console.error('Error creating/finding thread:', error);
-                    // If we can't create a thread, use the channel directly
-                    thread = null;
-                }
-            }
-        } else if (interaction.channel?.isThread()) {
-            thread = interaction.channel;
-        }
-
-        // Use consolidated check for existing search request
-        const existingRequest = await checkExistingSearchRequest(
-            thread || interaction.channel,
-            searchInfo.suggestedQuery,
-            interaction.client.user.id
-        );
-
-        if (existingRequest) {
-            completeSearch(channelId, searchInfo.suggestedQuery);
-            return `I've already requested a search for "${searchInfo.suggestedQuery}". Please approve or deny that request first.`;
-        }
-
-        // Request search permission
-        const requestData = await AISearchHandler.requestSearch(
-            interaction,
-            searchInfo.suggestedQuery,
-            searchInfo.reason
-        );
-
-        // If requestId is null, it means the search was executed automatically
-        // because approval is not required for this guild
-        if (requestData && requestData.requestId === null) {
-            completeSearch(channelId, searchInfo.suggestedQuery);
-
-            // Generate final response using the search result
-            const guildId = interaction.guild?.id;
-            const systemPrompt = await getPromptWithGuildPersonality(interaction.user.id, guildId);
-            const conversationHistory = [
-                { role: 'system', content: systemPrompt },
-                { role: 'user', content: trimmedMessage },
-                { role: 'system', content: `Here is relevant information to help answer the question: ${requestData.result}` }
-            ];
-
-            const responseContent = await aiService.chatText(conversationHistory, {
-                preset: 'chat',
-                max_tokens: 500
-            });
-
-            await interaction.editReply(responseContent);
-            return null; // Indicate no further action
-        }
-
-        // Approval required: requestSearch returned the request ID as a string
-        const requestId = typeof requestData === 'string' ? requestData : requestData?.requestId;
-
-        // Update the permission request message to use chunking if needed
-        const requestContent = `🔍 **Search Request**\n\nI'd like to gather some up-to-date information about:\n> ${searchInfo.suggestedQuery}\n\n**Reason:** ${searchInfo.reason}\n\nDo you approve this search?`;
-        
-        const messageChunks = chunkMessage(requestContent);
-        
-        // Send the permission request with buttons
-        const permissionMessage = await (thread || interaction.channel).send({
-            content: messageChunks[0], // First chunk
-            components: [{
-                type: 1,
-                components: [
-                    {
-                        type: 2,
-                        custom_id: `approve_search_${requestId}`,
-                        label: 'Approve Search',
-                        style: 3 // Green
-                    },
-                    {
-                        type: 2,
-                        custom_id: `deny_search_${requestId}`,
-                        label: 'Deny Search',
-                        style: 4 // Red
-                    }
-                ]
-            }]
-        });
-
-        // Send any additional chunks as follow-up messages
-        for (let i = 1; i < messageChunks.length; i++) {
-            await (thread || interaction.channel).send({
-                content: messageChunks[i]
-            });
-        }
-
-        // If we received a search response, try to store it in the database
-        const searchResponse = `🔍 I've requested permission to search for information about "${searchInfo.suggestedQuery}". Please approve or deny the request.`;
-        try {
-            const localBotUserId = getOrCreateUser(interaction.client.user.id, 'Goobster');
-
-            // Resolve the guild conversation row for this channel/thread
-            const localThreadId = thread?.id || createPlaceholderThreadId(interaction.channel?.id || interaction.channelId);
-            const guildConvRow = db.get(
-                `SELECT id FROM guild_conversations
-                 WHERE guildId = @guildId AND channelId = @channelId AND threadId = @threadId`,
-                {
-                    guildId: interaction.guildId,
-                    channelId: interaction.channel?.id || interaction.channelId,
-                    threadId: localThreadId
-                }
-            );
-
-            if (guildConvRow) {
-                const userRow = db.get('SELECT id FROM users WHERE discordId = @discordId', { discordId: interaction.user.id });
-                if (userRow) {
-                    const conversationRow = db.get(
-                        'SELECT id FROM conversations WHERE userId = @userId AND guildConversationId = @guildConvId',
-                        { userId: userRow.id, guildConvId: guildConvRow.id }
-                    );
-
-                    if (conversationRow) {
-                        db.run(
-                            `INSERT INTO messages (conversationId, guildConversationId, createdBy, message, isBot, metadata)
-                             VALUES (@conversationId, @guildConvId, @createdBy, @message, 1, @metadata)`,
-                            {
-                                conversationId: conversationRow.id,
-                                guildConvId: guildConvRow.id,
-                                createdBy: localBotUserId,
-                                message: searchResponse,
-                                metadata: JSON.stringify({ pendingSearch: true, query: searchInfo.suggestedQuery })
-                            }
-                        );
-                    }
-                }
-            }
-        } catch (dbError) {
-            console.error('Database error during search processing:', dbError);
-            // Continue with the search response even if DB operations fail
-        }
-
-        return searchResponse;
-    } catch (error) {
-        // Make sure to remove the search tracking on error
-        completeSearch(channelId, searchInfo.suggestedQuery);
-        // Re-throw the error to be handled by the caller
-        throw error;
-    }
-}
+const { DEFAULT_PROMPT } = require('./chat/prompts');
+const {
+    logSystemEvent,
+    getOrCreateUser,
+    getOrCreateConversation,
+    checkDatabaseHealth,
+    diagnoseDatabaseIssues,
+    createPlaceholderThreadId,
+    trackMessage
+} = require('./chat/chatDb');
+const { getContextWithSummary } = require('./chat/chatContext');
+const {
+    pendingImageGenerations,
+    handleAIResponse,
+    detectSearchNeed,
+    handleSearchFlow
+} = require('./chat/searchFlow');
+const { sendChunkedResponse } = require('./chat/responder');
+const { handleReactionAdd, handleReactionRemove } = require('./chat/reactions');
+const { getThreadName } = require('./chat/threadManager');
 
 async function handleChatInteraction(interaction, thread = null) {
     let conversationId = null;
@@ -999,12 +141,6 @@ async function handleChatInteraction(interaction, thread = null) {
             console.log('Handling role-style bot mention as a direct mention');
         }
 
-        // Get thread preference for this guild/conversation - safely handle missing guild
-        let threadPreference = THREAD_PREFERENCE.ALWAYS_CHANNEL; // Default
-        if (interaction.guildId) {
-            threadPreference = await getThreadPreference(interaction.guildId);
-        }
-        
         // Thread usage disabled – always converse in the channel
         thread = null;
 
@@ -1466,7 +602,7 @@ This directive applies only in this server and overrides any conflicting instruc
 
             // For non-image responses, use existing chunking
             const chunks = chunkMessage(processedResponse);
-            await sendChunkedResponse(interaction, chunks, false, thread);
+            await sendChunkedResponse(interaction, chunks, false);
             userResponseSent = true; // Mark that we've sent a response
             
             // Store messages in database with transaction
@@ -1577,365 +713,9 @@ This directive applies only in this server and overrides any conflicting instruc
     }
 }
 
-// Add helper function for sending chunked responses
-async function sendChunkedResponse(interaction, chunks, isError = false) {
-    try {
-        // Use existing thread if provided, otherwise check thread preference
-        let thread = interaction.channel;
-        
-        // Only check thread preference if no thread is provided and we're not already in a thread
-        if (!thread && !interaction.channel?.isThread() && interaction.guildId) {
-            // Get the guild's thread preference
-            const threadPreference = await getThreadPreference(interaction.guildId);
-            
-            // If preference is ALWAYS_THREAD, create/use a thread
-            if (threadPreference === THREAD_PREFERENCE.ALWAYS_THREAD) {
-                const channelName = interaction.channel.name.replace(/[^a-z0-9-_]/gi, '-').toLowerCase();
-                const threadName = `goobster-chat-${channelName}`;
-                
-                try {
-                    thread = await getOrCreateThreadSafely(interaction.channel, threadName);
-                    
-                    // Send welcome message only for newly created threads
-                    if (!thread.messages.cache.size) {
-                        await thread.send(
-                            "👋 Hi! I've created this thread for our conversation. " +
-                            "You can continue chatting with me here by:\n" +
-                            "1. Using `/chat` command\n" +
-                            `2. Mentioning me (@${interaction.client.user.username})\n\n` +
-                            "The thread will keep our conversation organized and maintain context!"
-                        );
-                    }
-                } catch (error) {
-                    console.error('Error creating/finding thread:', error);
-                    // If we can't create a thread, use the channel directly
-                    thread = null;
-                }
-            }
-        } else if (interaction.channel?.isThread()) {
-            thread = interaction.channel;
-        }
-
-        // For slash commands, use the interaction reply mechanism
-        if (interaction.commandName === 'chat') {
-            // Send first chunk as reply or edit
-            if (interaction.deferred || interaction.replied) {
-                await interaction.editReply(chunks[0]);
-            } else {
-                await interaction.reply(chunks[0]);
-            }
-
-            // Send remaining chunks as follow-ups
-            for (let i = 1; i < chunks.length; i++) {
-                await interaction.followUp(chunks[i]);
-            }
-            
-            // If we created a thread, send a message to direct the user there
-            if (thread && !interaction.channel.isThread()) {
-                await interaction.followUp({
-                    content: `I've continued our conversation in a thread: ${thread}`,
-                    ephemeral: true
-                });
-            }
-        } 
-        // For mentions, send to the appropriate channel based on thread preference
-        else {
-            const targetChannel = thread || interaction.channel;
-            
-            // Send all chunks to the target channel
-            for (const chunk of chunks) {
-                await targetChannel.send(chunk);
-            }
-            
-            // If we created a thread, send a message to direct the user there
-            if (thread && interaction.channel !== thread) {
-                await interaction.channel.send({
-                    content: `I've continued our conversation in a thread: ${thread}`,
-                    allowedMentions: { users: [], roles: [] }
-                });
-            }
-        }
-    } catch (error) {
-        console.error('Error sending chunked response:', {
-            error: error.message,
-            stack: error.stack,
-            isErrorResponse: isError
-        });
-        
-        // If this is already an error response, don't try again
-        if (!isError) {
-            const errorChunks = chunkMessage("I encountered an error while sending my response. Please try again.");
-            await sendChunkedResponse(interaction, errorChunks, true);
-        }
-    }
-}
-
-// Add reaction handler for conversation branching
-async function handleReactionAdd(reaction, user) {
-    console.log('Reaction add triggered:', {
-        emoji: reaction.emoji.name,
-        user: user.tag,
-        messageId: reaction.message.id,
-        channelId: reaction.message.channel.id
-    });
-
-    if (user.bot) {
-        console.log('Ignoring bot reaction');
-        return;
-    }
-
-    const msg = reaction.message;
-    console.log('Processing reaction:', reaction.emoji.name);
-
-    try {
-        if (reaction.emoji.name === '🔄') {
-            console.log('Handling regenerate reaction');
-            await msg.channel.sendTyping();
-            
-            try {
-                // Find the user's message that triggered this response
-                const messages = await msg.channel.messages.fetch({ limit: 50, before: msg.id });
-                const userMessage = messages.find(m => !m.author.bot && m.content);
-                
-                if (!userMessage) {
-                    await msg.reply("I couldn't find the original message to regenerate a response for.");
-                    return;
-                }
-
-                // Get the current prompt (falls back to the default prompt)
-                const promptRow = db.get(
-                    `SELECT p.prompt
-                     FROM prompts p
-                     JOIN guild_conversations gc ON gc.promptId = p.id
-                     WHERE gc.threadId = @threadId`,
-                    { threadId: msg.channel.id }
-                );
-
-                // Create a new completion with slightly higher temperature for variety
-                const newResponse = await aiService.chatText([
-                        { role: 'system', content: promptRow?.prompt || DEFAULT_PROMPT },
-                        { role: 'user', content: userMessage.content }
-                    ], {
-                        preset: 'creative',
-                        max_tokens: 500
-                    });
-
-                // Send the new response
-                const response = await msg.reply({
-                    content: `🔄 **Regenerated Response:**\n\n${newResponse}`,
-                    allowedMentions: { users: [], roles: [] }
-                });
-
-                // Add the standard reaction controls
-                await response.react('🔄');
-                await response.react('📌');
-                await response.react('🌳');
-                await response.react('💡');
-                await response.react('🔍');
-                await response.react('📝');
-
-            } catch (error) {
-                console.error('Error in response regeneration:', error);
-                await msg.reply("I encountered an error while regenerating the response. Please try again.");
-            }
-        } else if (reaction.emoji.name === '📌') {
-            console.log('Handling pin reaction');
-            try {
-                await msg.pin();
-                await msg.react('📍');
-            } catch (pinError) {
-                console.error('Error pinning message:', pinError);
-                throw pinError;
-            }
-        } else if (reaction.emoji.name === '🌱') {
-            // Create descriptive branch name from message content
-            const branchTopic = msg.content
-                .split(/[.!?]/)[0]  // Get first sentence
-                .slice(0, 30)       // Take first 30 chars
-                .trim()
-                .toLowerCase()
-                .replace(/[^a-z0-9]+/g, '-');  // Convert to URL-friendly format
-
-            const branchName = `branch-${branchTopic}-${msg.id.slice(-4)}`;
-            
-            // Create conversation branch
-            const newThread = await msg.channel.threads.create({
-                name: branchName,
-                startMessage: msg,
-                autoArchiveDuration: ThreadAutoArchiveDuration.OneWeek
-            });
-            await newThread.send("🌱 New conversation branch created! Previous context will be maintained.");
-        } else if (reaction.emoji.name === '💡') {
-            // Mark as solution/helpful
-            await msg.react('✨');
-            await msg.reply("Marked as helpful solution! 💡");
-        } else if (reaction.emoji.name === '🔍') {
-            // Expand on this topic
-            await msg.channel.sendTyping();
-            
-            // Create deep-dive prompt
-            const deepDivePrompt = [
-                { role: 'system', content: 'You are helping to expand on a previous response. Provide more detailed information, examples, and explanations about the topic. Be thorough but maintain clarity. Structure your response with clear sections using markdown headers.' },
-                { role: 'user', content: `Please provide a detailed explanation and expansion of this topic: "${msg.content}"` }
-            ];
-
-            try {
-                const expandedResponse = await aiService.chatText(deepDivePrompt, {
-                    preset: 'chat',
-                    max_tokens: 1000
-                });
-
-                // Use the chunked reply utility
-                const chunks = chunkMessage(expandedResponse);
-                for (const chunk of chunks) {
-                    await msg.channel.send(chunk);
-                }
-            } catch (error) {
-                console.error('Error in deep-dive generation:', error);
-                await msg.reply("I encountered an error while generating the detailed explanation. Please try again.");
-            }
-        } else if (reaction.emoji.name === '📝') {
-            // Request summary of thread up to this point
-            await msg.channel.sendTyping();
-            
-            try {
-                // Fetch messages up to this point
-                const messages = await msg.channel.messages.fetch({ 
-                    limit: 100,
-                    before: msg.id 
-                });
-                
-                // Filter and format messages
-                const conversationText = messages
-                    .reverse()
-                    .map(m => `${m.author.username}: ${m.content}`)
-                    .join('\n');
-
-                const summaryPrompt = [
-                    { role: 'system', content: 'Create a concise but comprehensive summary of the conversation. Focus on key points, decisions, and important information. Use bullet points for clarity.' },
-                    { role: 'user', content: `Please summarize this conversation:\n\n${conversationText}` }
-                ];
-
-                const summary = await aiService.chatText(
-                    summaryPrompt,
-                    {
-                        temperature: 0.7,
-                        max_tokens: 500
-                    }
-                );
-
-                const response = await msg.reply({
-                    content: `📝 **Conversation Summary:**\n\n${summary}`,
-                    allowedMentions: { users: [], roles: [] }
-                });
-
-                // Add pin reaction for easy reference
-                await response.react('📌');
-            } catch (error) {
-                console.error('Error generating summary:', error);
-                await msg.reply("I encountered an error while generating the summary. Please try again.");
-            }
-        }
-    } catch (error) {
-        console.error('Error in handleReactionAdd:', error);
-        // Try to notify the user of the error
-        try {
-            await msg.channel.send(`Error processing reaction: ${error.message}`);
-        } catch (notifyError) {
-            console.error('Could not notify user of error:', notifyError);
-        }
-    }
-}
-
-async function handleReactionRemove(reaction, user) {
-    if (user.bot) return;
-
-    const msg = reaction.message;
-    if (reaction.emoji.name === '📌') {
-        // Unpin message if no 📌 reactions remain
-        const pinReactions = msg.reactions.cache.get('📌');
-        if (!pinReactions || pinReactions.count === 0) {
-            try {
-                await msg.unpin();
-                // Remove the pin confirmation reaction if it exists
-                const confirmReaction = msg.reactions.cache.get('📍');
-                if (confirmReaction) {
-                    await confirmReaction.remove();
-                }
-            } catch (error) {
-                console.error('Error unpinning message:', error);
-            }
-        }
-    }
-    // Add other reaction removal handlers as needed
-}
-
-// Add this new function
-async function getOrCreateThreadSafely(channel, threadName) {
-    // If the channel is already a thread, just return it
-    if (channel.isThread()) {
-        return channel;
-    }
-    
-    const lockKey = `${channel.id}-${threadName}`;
-    if (threadLocks.has(lockKey)) {
-        return await threadLocks.get(lockKey);
-    }
-
-    const lockPromise = (async () => {
-        try {
-            const threads = await channel.threads.fetch();
-            let thread = threads.threads.find(t => t.name === threadName);
-            
-            if (!thread) {
-                console.log(`Creating new thread "${threadName}" in channel ${channel.name}`);
-                thread = await channel.threads.create({
-                    name: threadName,
-                    autoArchiveDuration: ThreadAutoArchiveDuration.OneWeek,
-                    reason: 'New Goobster chat thread'
-                });
-            } else {
-                console.log(`Found existing thread "${threadName}" in channel ${channel.name}`);
-            }
-
-            // Make sure thread is unarchived
-            if (thread.archived) {
-                console.log(`Unarchiving thread "${threadName}"`);
-                await thread.setArchived(false);
-            }
-
-            return thread;
-        } catch (error) {
-            console.error('Error in thread creation:', error);
-            // Return the original channel as fallback
-            return channel;
-        }
-    })();
-
-    threadLocks.set(lockKey, lockPromise);
-    try {
-        return await lockPromise;
-    } finally {
-        threadLocks.delete(lockKey);
-    }
-}
-
-// TODO: Add proper handling for message context management
-// TODO: Add proper handling for message summary failures
-// TODO: Add proper handling for message chunking failures
-// TODO: Add proper handling for message fetch failures
-// TODO: Add proper handling for message cleanup
-// TODO: Add proper handling for message reference resolution
-// TODO: Add proper handling for thread state management
-// TODO: Add proper handling for thread creation failures
-// TODO: Add proper handling for thread archival
-// TODO: Add proper handling for thread lock timeouts
-
 // Update the function that processes messages to incorporate guild personality directives
 async function processMessage(message, isThread = false) {
     try {
-        // ... existing code ...
-
         // Get guild ID for guild-specific settings
         const guildId = message.guild?.id;
         
@@ -1955,9 +735,6 @@ ${personalityDirective}
 This directive applies only in this server and overrides any conflicting instructions.`;
             }
         }
-        
-        // Use the enhanced prompt for the message processing
-        // ... continue with message processing using systemPrompt ...
 
         // Add image detection
         const imageRequest = await imageDetectionHandler.detectImageGenerationRequest(message.content);
@@ -2007,86 +784,8 @@ This directive applies only in this server and overrides any conflicting instruc
                 return;
             }
         }
-
-        // Continue with existing message processing...
     } catch (error) {
         console.error('Error processing message:', error);
-        // ... existing error handling ...
-    }
-}
-
-/**
- * Tracks a message in the conversation history
- * @param {string} guildConvId - The guild conversation ID
- * @param {string} discordUserId - The Discord user ID
- * @param {string} message - The message content
- * @param {string} role - The role ('user' or 'assistant')
- */
-async function trackMessage(guildConvId, discordUserId, message, role) {
-    try {
-        const userId = getOrCreateUser(discordUserId, `user_${discordUserId}`);
-        const conversationId = getOrCreateConversation(userId, guildConvId);
-
-        db.run(
-            `INSERT INTO messages (conversationId, guildConversationId, createdBy, message, isBot)
-             VALUES (@conversationId, @guildConvId, @createdBy, @message, @isBot)`,
-            {
-                conversationId,
-                guildConvId,
-                createdBy: userId,
-                message,
-                isBot: role === 'assistant'
-            }
-        );
-    } catch (error) {
-        console.error('Error tracking message in database:', {
-            error: error.message,
-            stack: error.stack,
-            guildConvId,
-            discordUserId,
-            messagePreview: message.substring(0, 50) + (message.length > 50 ? '...' : ''),
-            role
-        });
-        // Don't throw the error, just log it - we don't want to interrupt the flow
-    }
-}
-
-/**
- * Generate a thread name based on the user
- * @param {Object} user - The Discord user object
- * @returns {string} - A thread name
- */
-async function getThreadName(user) {
-    try {
-        // Generate thread name using OpenAI
-        const prompt = `
-Generate a short, creative, and friendly thread name for a conversation with a user named ${user.username}.
-The name should be related to having a chat or conversation in a fun way.
-Keep it under 30 characters (including spaces) and make it appropriate for all ages.
-Return ONLY the thread name without any quotation marks or additional text.
-`;
-
-        let threadName = (await aiService.chatText([
-            { role: 'user', content: prompt }
-        ], {
-            preset: 'chat',
-            max_tokens: 20
-        })).trim();
-        
-        // Ensure thread name meets Discord requirements
-        if (threadName.length > 100) {
-            threadName = threadName.substring(0, 97) + '...';
-        }
-        
-        // Fall back to a basic name if generation fails or is empty
-        if (!threadName) {
-            threadName = `Chat with ${user.username}`;
-        }
-        
-        return threadName;
-    } catch (error) {
-        console.error('Error generating thread name:', error);
-        return `Chat with ${user.username}`;
     }
 }
 
@@ -2100,4 +799,4 @@ module.exports = {
     diagnoseDatabaseIssues,
     checkDatabaseHealth,
     logSystemEvent
-}; 
+};
